refactor(core): extract LLM telemetry into execution::llm_telemetry
Move the four LLM span/metric stampers (stamp_llm_span_cost_attribution,
stamp_llm_runtime_backend, insert_llm_streaming_metrics,
mirror_llm_metrics_to_span) plus their 6-test submodule (GLOBAL_TRACE_LOCK,
WireLabelStub, capture helpers) off the executor into a new feature-gated
execution::llm_telemetry module. The module is gated at its `mod`
declaration so individual #[cfg]s collapse; a cfg-gated `use` keeps the 16
executor call sites unchanged. mark_execution_terminal stays (public-API
silent-guard contract); the outer-span inline stamp in execute_impl is
untouched (distinct span, not a duplicate).

Also drop the executor's duplicate test_chat_template_format_from_str —
chat_template.rs::test_format_from_str is a strict superset.

No behavior change; executor.rs drops ~406 lines.
